+++ back/app/services/idea_custom_fields_service.rb
@@ -1,13 +1,52 @@
 # frozen_string_literal: true
 
 class IdeaCustomFieldsService
-  def all_fields(custom_form, _options = {})
-    default_fields(custom_form)
+  def initialize(custom_form)
+    @custom_form = custom_form
+  end
+
+  def all_fields
+    default_fields
+  end
+
+  def configurable_fields
+    disallowed_fields = %w[author_id budget]
+    all_fields.reject do |field|
+      disallowed_fields.include? field.code
+    end
+  end
+
+  def reportable_fields
+    enabled_fields.reject(&:built_in?)
+  end
+
+  def visible_fields
+    enabled_fields
+  end
+
+  def enabled_fields
+    all_fields.select(&:enabled?)
+  end
+
+  def extra_visible_fields
+    visible_fields.reject(&:built_in?)
+  end
+
+  def allowed_extra_field_keys
+    fields_with_array_keys, fields_with_simple_keys = extra_visible_fields.partition do |field|
+      field.input_type == 'multiselect'
+    end
+    [
+      *fields_with_simple_keys.map(&:key).map(&:to_sym),
+      fields_with_array_keys.map(&:key).map(&:to_sym).index_with { |_k| [] }
+    ]
   end
 
   private
 
-  def default_fields(custom_form)
+  attr_reader :custom_form
+
+  def default_fields
     ml_s = MultilocService.new
     [
       CustomField.new(
@@ -30,7 +69,7 @@ class IdeaCustomFieldsService
         end,
         required: true,
         enabled: true,
-        ordering: 1
+        ordering: 0
       ),
       CustomField.new(
         id: SecureRandom.uuid,
@@ -52,7 +91,7 @@ class IdeaCustomFieldsService
         end,
         required: true,
         enabled: true,
-        ordering: 2
+        ordering: 1
       ),
       CustomField.new(
         id: SecureRandom.uuid,
@@ -74,7 +113,7 @@ class IdeaCustomFieldsService
         end,
         required: false,
         enabled: true,
-        ordering: 3
+        ordering: 2
       ),
       CustomField.new(
         id: SecureRandom.uuid,
@@ -96,7 +135,7 @@ class IdeaCustomFieldsService
         end,
         required: false,
         enabled: true,
-        ordering: 4
+        ordering: 3
       ),
       CustomField.new(
         id: SecureRandom.uuid,
@@ -167,29 +206,6 @@ class IdeaCustomFieldsService
       CustomField.new(
         id: SecureRandom.uuid,
         resource: custom_form,
-        key: 'location_point_geojson',
-        code: 'location_point_geojson',
-        input_type: 'point',
-        title_multiloc: ml_s.i18n_to_multiloc(
-          'custom_fields.ideas.location.title',
-          locales: CL2_SUPPORTED_LOCALES
-        ),
-        description_multiloc: begin
-          ml_s.i18n_to_multiloc(
-            'custom_fields.ideas.location.description',
-            locales: CL2_SUPPORTED_LOCALES
-          )
-        rescue StandardError
-          {}
-        end,
-        required: false,
-        hidden: true,
-        enabled: true,
-        ordering: 7
-      ),
-      CustomField.new(
-        id: SecureRandom.uuid,
-        resource: custom_form,
         key: 'idea_images_attributes',
         code: 'idea_images_attributes',
         input_type: 'image_files',
@@ -207,7 +223,7 @@ class IdeaCustomFieldsService
         end,
         required: false,
         enabled: true,
-        ordering: 8
+        ordering: 7
       ),
       CustomField.new(
         id: SecureRandom.uuid,
@@ -229,7 +245,7 @@ class IdeaCustomFieldsService
         end,
         required: false,
         enabled: true,
-        ordering: 9
+        ordering: 8
       )
     ]
   end